+++ axum/src/extract/mod.rs
@@ -12,6 +12,7 @@ pub mod rejection;
 pub mod ws;
 
 mod content_length_limit;
+mod host;
 mod raw_query;
 mod request_parts;
 
@@ -23,6 +24,7 @@ pub use self::{
     connect_info::ConnectInfo,
     content_length_limit::ContentLengthLimit,
     extractor_middleware::extractor_middleware,
+    host::Host,
     path::Path,
     raw_query::RawQuery,
     request_parts::{BodyStream, RawBody},